x86-64: mingw: Return _Float128 in xmm0 [PR115054]
For MinGW on x86-64, GCC currently both passes and returns `_Float128` on the stack.
This differs from Clang's behavior of returning in xmm0.

The same as


There is no official ABI from Microsoft so any calling convention is
an extension.

Currently a mismatch with Clang so a change on at least one side
is needed.

There is no official ABI from Microsoft so this is purely an extension.

Thus, update `TFmode` to return in xmm0

Must be passed indirectly because the ABI requires a shadow slot

2026-03-25  Trevor Gross  <tmgross@umich.edu>

gcc/:
	PR target/115054
	* gcc/config/i386/i386.cc (function_value_ms_64): Return
	_Float128 in xmm0 on Windows.

Signed-off-by: Trevor Gross <tmgross@umich.edu>